made it possible to pick more images
