+++ lib/Redesign/Testin_components/showcase/showcase.dart
@@ -81,7 +81,6 @@ class ShowcasePage extends StatelessWidget {
               Column(
                 children: [
                   MainImagePlusBoxes(),
-                  CreateMainImagePlusBoxes(),
                 ],
               ),
               Column(
